back-end/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-environ
psycopg2-binary
pandas
mod_wsgi
imblearn
scikit-learn
joblib